# server/openapi.yaml
openapi: 3.0.3
info:
  title: JXC Server API
  description: |
    极速云进销存（JXC）服务端 OpenAPI 3.0 初版。

    - 统一响应：`code/message/data/request_id`
    - 鉴权：`Authorization: Bearer <access_token>`
    - 写接口幂等：`X-Idempotency-Key`
    - 金额字段：字符串传输
  version: 1.1.0
servers:
  - url: http://127.0.0.1:8080
    description: Local Development
...
paths:
...
  /api/v1/inventory/stock-checks:
    post:
      tags: [StockChecks]
      summary: 创建盘点单
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockCheckCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessJson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestId:
      name: x-request-id
      in: header
      required: false
      schema:
        type: string
      description: 请求链路 ID（可选，不传则服务端自动生成）
    XIdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
      description: 写接口幂等键（必填）
...
  headers:
    XRequestIdHeader:
      description: 服务端返回的请求链路 ID
      schema:
        type: string
    XIdempotentReplayHeader:
      description: 幂等重放标识（true 表示重放）
      schema:
        type: string
  responses:
...
    SuccessJson:
      description: 业务成功
      headers:
        x-request-id:
          $ref: '#/components/headers/XRequestIdHeader'
        x-idempotent-replay:
          $ref: '#/components/headers/XIdempotentReplayHeader'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiSuccess'
...
    BadRequest:
      description: 参数错误（业务码常见 4000）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: 未登录或 Token 无效（业务码常见 4010）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: 无权限（业务码常见 4030）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Conflict:
      description: 冲突（业务码常见 4090/4091/4092）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiSuccess:
      type: object
      required: [code, message, data, request_id]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          additionalProperties: true
          description: 不同接口返回结构不同
        request_id:
          type: string
          example: req_20260303_xxx
    ApiError:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: integer
          example: 4000
        message:
          type: string
          example: 参数错误
        data:
          type: object
          nullable: true
          additionalProperties: true
        request_id:
          type: string
          example: req_20260303_xxx
...
    StockCheckCreateItemRequest:
      type: object
      required: [product_id]
      properties:
        product_id:
          type: integer
          format: int64
    StockCheckCreateRequest:
      type: object
      required: [biz_no, items]
      properties:
        biz_no:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/StockCheckCreateItemRequest'
        remark:
          type: string
...